Add files via upload
To better tailor the insurance package to be given to different customers, the company now wants to integrate machine learning into the process and identify the right insurance charges for each customer. You are hired as a ML Engineer to help the company predict the right charges based on the data that they have been maintaining for each customer.